extracted event distribution strategy and added Pipe concept
I still don't think there is a single event distribution strategy which
should be reified by including it in the a general base class. Further
than that I see a lot of value in the simplicity of a single event handler
per source with the option of explicit fanout strategies where appropriate.

While implementing this I noticed that "onEvent(Event)" and "raise(Event)"
are just ways of saying the same thing, so I unified them. This in turn
helped to simplify some of the basic event passing components. For conciseness I
invented a sub-interface "EventPipe" which is both source and handler and
interposed a base class under the filters